feat: Typescript conversion
* Convert codebase to typescript for better DX
