+++ jest.config.js
@@ -1,0 +1,5 @@
+module.exports = {
+    roots: ['<rootDir>/src'],
+    testRegex: '(/__tests__/.*|(\\.|/)(test|spec))\\.(ts|tsx?)$',
+    moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx', 'json', 'node'],
+}
