--- server/routes/api/ApiRouter.js
@@ -1,9 +1,0 @@
-const { Router } = require('express');
-
-const router = new Router();
-
-router.get("/", (_res, res) => {
-  res.json({ message: 'Server running!'})
-});
-
-module.exports = router;
