+++ server/routes/api/ApiRouter.test.ts
@@ -1,6 +1,6 @@
-const express = require('express'); // import express
-const apiRouter = require("./ApiRouter"); //import file we are testing
-const request = require("supertest"); // supertest is a framework that allows to easily test web apis
+import * as express from "express";   // import express
+import apiRouter from "./ApiRouter";  // import file we are testing
+import request from "supertest";      // supertest is a framework that allows to easily test web apis
 
 const app = express(); //an instance of an express app, a 'fake' express app
 app.use("/api", apiRouter); //routes
@@ -10,7 +10,7 @@ describe('server', () => {
     it('should return success from /api GET', async () => {
       const { body, status } = await request(app).get("/api");
       expect(status).toEqual(200);
-      expect(body).toEqual({ message: 'Server running!'})
+      expect(body).toEqual({ message: 'Server running!' })
     });
   });
 })
