--- server/routes/api/index.js
@@ -1,5 +1,0 @@
-const apiRouter = require('./ApiRouter');
-
-module.exports = {
-  apiRouter,
-}
