+++ server/routes/api/index.ts
@@ -1,0 +1,5 @@
+import apiRouter from './ApiRouter';
+
+export {
+  apiRouter,
+}
