--- server/routes/events/index.js
@@ -1,5 +1,0 @@
-const eventsRouter = require("./EventsRouter")
-
-module.exports = {
-  eventsRouter
-}
