+++ server/routes/events/index.ts
@@ -1,0 +1,5 @@
+import eventsRouter from "./EventsRouter"
+
+export {
+  eventsRouter
+}
