--- server/services/seeds/SeedsService.js
@@ -1,9 +1,0 @@
-const seedsData = require('../../database/SeedsReposiory');
-
-const SeedsService = () => ({
-  getAll() {
-    return seedsData.getAll();
-  }
-});
-
-module.exports = SeedsService();
