+++ server/services/seeds/SeedsService.ts
@@ -1,0 +1,9 @@
+import seedsData from '../../database/SeedsRepository';
+
+const SeedsService = () => ({
+  getAll() {
+    return seedsData.getAll();
+  }
+});
+
+export default SeedsService();
